ingest-api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.9
psycopg2-binary>=2.9
sentence-transformers>=3.0
tiktoken>=0.7
click>=8.1
python-docx>=1.1
beautifulsoup4>=4.12
pdfplumber>=0.11
pymupdf>=1.24
